+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 ## 1.4 branch
 # 1.5.0
 * the plugin has been migrated to CakePHP 3.6;
-* removed `Thumber\ThumbTrait::getUrl()` method;
+* removed `getExtension()` and `getUrl()` methods provided by the
+    `Thumber\ThumbTrait` trait class;
 * removed `Thumber\ThumbTrait::getSupportedFormats()` method, added 
     `Thumber\Utility\ThumbManager::$supportedFormats` static property;
 * `Cake\Network\Exception\InternalErrorException` exception has been replaced